Refactor GymPhotosDialog component
Refactored the `GymPhotosDialog` component into smaller components to improve maintainability.  The original component was exceeding 250 lines of code.
